refactor: moved options to separate module
